Move getDates to utils
